+++ variables.tf
@@ -25,28 +25,32 @@ variable "number_of_instances" {
 }
 
 variable "use_num_suffix" {
-  description = "Always append numerical suffix(like 001, 002 and so on) to instance name and host name, even if number_of_instances is 1"
+  description = "Always append numerical suffix(like 001, 002 and so on) to instance name and host name, even if number_of_instances is 1."
   type        = bool
   default     = false
 }
 
 variable "image_id" {
   description = "The image id used to launch one or more ecs instances."
+  type        = string
   default     = ""
 }
 
 variable "image_ids" {
   description = "A list of ecs image IDs to launch one or more ecs instances."
+  type        = string
   default     = []
 }
 
 variable "instance_type" {
   description = "The instance type used to launch one or more ecs instances."
+  type        = string
   default     = ""
 }
 
 variable "credit_specification" {
   description = "Performance mode of the t5 burstable instance. Valid values: 'Standard', 'Unlimited'."
+  type        = string
   default     = ""
 }
 
@@ -58,6 +62,7 @@ variable "security_group_ids" {
 
 variable "name" {
   description = "Name to be used on all resources as prefix. Default to 'TF-Module-ECS-Instance'. The final default name would be TF-Module-ECS-Instance001, TF-Module-ECS-Instance002 and so on."
+  type        = string
   default     = ""
 }
 
@@ -68,37 +73,43 @@ variable "description" {
 }
 variable "resource_group_id" {
   description = "The Id of resource group which the instance belongs."
+  type        = string
   default     = ""
 }
 
 variable "internet_charge_type" {
   description = "The internet charge type of instance. Choices are 'PayByTraffic' and 'PayByBandwidth'."
+  type        = string
   default     = "PayByTraffic"
 }
 
 variable "host_name" {
   description = "Host name used on all instances as prefix. Like if the value is TF-ECS-Host-Name and then the final host name would be TF-ECS-Host-Name001, TF-ECS-Host-Name002 and so on."
+  type        = string
   default     = ""
 }
 
 variable "password" {
   description = "The password of instance."
+  type        = string
   default     = ""
 }
 
 variable "kms_encrypted_password" {
-  description = "An KMS encrypts password used to an instance. It is conflicted with `password`."
+  description = "An KMS encrypts password used to an instance. It is conflicted with 'password'."
+  type        = string
   default     = ""
 }
 
 variable "kms_encryption_context" {
-  description = "An KMS encryption context used to decrypt `kms_encrypted_password` before creating or updating an instance with `kms_encrypted_password`"
+  description = "An KMS encryption context used to decrypt 'kms_encrypted_password' before creating or updating an instance with 'kms_encrypted_password'."
   type        = map(string)
   default     = {}
 }
 
 variable "system_disk_category" {
   description = "The system disk category used to launch one or more ecs instances."
+  type        = string
   default     = "cloud_efficiency"
 }
 
@@ -108,14 +119,21 @@ variable "system_disk_size" {
   default     = 40
 }
 
+variable "system_disk_auto_snapshot_policy_id" {
+  description = "The ID of the automatic snapshot policy applied to the system disk."
+  type        = string
+  default     = ""
+}
+
 variable "data_disks" {
-  description = "Additional data disks to attach to the scaled ECS instance"
+  description = "Additional data disks to attach to the scaled ECS instance."
   type        = list(map(string))
   default     = []
 }
 
 variable "vswitch_id" {
   description = "The virtual switch ID to launch in VPC."
+  type        = string
   default     = ""
 }
 
@@ -126,7 +144,8 @@ variable "vswitch_ids" {
 }
 
 variable "private_ip" {
-  description = "Configure Instance private IP address"
+  description = "Configure Instance private IP address."
+  type        = string
   default     = ""
 }
 
@@ -156,6 +175,7 @@ variable "associate_public_ip_address" {
 
 variable "instance_charge_type" {
   description = "The charge type of instance. Choices are 'PostPaid' and 'PrePaid'."
+  type        = string
   default     = "PostPaid"
 }
 
@@ -166,27 +186,32 @@ variable "dry_run" {
 }
 
 variable "user_data" {
-  description = "User data to pass to instance on boot"
+  description = "User data to pass to instance on boot."
+  type        = string
   default     = ""
 }
 
 variable "role_name" {
-  description = "Instance RAM role name. The name is provided and maintained by RAM. You can use `alicloud_ram_role` to create a new one."
+  description = "Instance RAM role name. The name is provided and maintained by RAM. You can use 'alicloud_ram_role' to create a new one."
+  type        = string
   default     = ""
 }
 
 variable "key_name" {
   description = "The name of SSH key pair that can login ECS instance successfully without password. If it is specified, the password would be invalid."
+  type        = string
   default     = ""
 }
 
 variable "spot_strategy" {
-  description = "The spot strategy of a Pay-As-You-Go instance, and it takes effect only when parameter `instance_charge_type` is 'PostPaid'. Value range: 'NoSpot': A regular Pay-As-You-Go instance. 'SpotWithPriceLimit': A price threshold for a spot instance. 'SpotAsPriceGo': A price that is based on the highest Pay-As-You-Go instance"
+  description = "The spot strategy of a Pay-As-You-Go instance, and it takes effect only when parameter 'instance_charge_type' is 'PostPaid'. Value range: 'NoSpot': A regular Pay-As-You-Go instance. 'SpotWithPriceLimit': A price threshold for a spot instance. 'SpotAsPriceGo': A price that is based on the highest Pay-As-You-Go instance."
+  type        = string
   default     = "NoSpot"
 }
 
 variable "spot_price_limit" {
   description = "The hourly price threshold of a instance, and it takes effect only when parameter 'spot_strategy' is 'SpotWithPriceLimit'. Three decimals is allowed at most."
+  type        = number
   default     = 0
 }
 
@@ -197,12 +222,14 @@ variable "deletion_protection" {
 }
 
 variable "force_delete" {
-  description = "If it is true, the 'PrePaid' instance will be change to 'PostPaid' and then deleted forcibly. However, because of changing instance charge type has CPU core count quota limitation, so strongly recommand that `Don't modify instance charge type frequentlly in one month`."
+  description = "If it is true, the 'PrePaid' instance will be change to 'PostPaid' and then deleted forcibly. However, because of changing instance charge type has CPU core count quota limitation, so strongly recommand that 'Don't modify instance charge type frequentlly in one month'."
+  type        = bool
   default     = false
 }
 
 variable "security_enhancement_strategy" {
   description = "The security enhancement strategy."
+  type        = string
   default     = "Active"
 }
 
@@ -232,38 +259,38 @@ variable "volume_tags" {
 
 # Depreceted parameters
 variable "instance_name" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `name` instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'name' instead."
   default     = ""
 }
 variable "group_ids" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `security_group_ids` instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'security_group_ids' instead."
   type        = list(string)
   default     = []
 }
 
 variable "system_category" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `system_disk_category` instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'system_disk_category' instead."
   default     = "cloud_efficiency"
 }
 
 variable "system_size" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `system_disk_size` replaces it."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'system_disk_size' replaces it."
   type        = number
   default     = 40
 }
 
 variable "disk_name" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `data_disks` 'name' instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'data_disks' 'name' instead."
   default     = "TF_ECS_Disk"
 }
 
 variable "disk_category" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `data_disks` 'category' instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'data_disks' 'category' instead."
   default     = "cloud_efficiency"
 }
 
 variable "disk_size" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `data_disks` 'size' instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'data_disks' 'size' instead."
   type        = number
   default     = 40
 }
@@ -274,7 +301,7 @@ variable "disk_tags" {
 
   default = {
     created_by   = "Terraform"
-    created_from = "module-tf-alicloud-ecs-instance"
+    created_from = "module-tf-alicloud-ecs-instance."
   }
 }
 variable "instance_tags" {
@@ -286,7 +313,7 @@ variable "instance_tags" {
   }
 }
 variable "number_of_disks" {
-  description = "(Deprecated) It has been deprecated from version 2.0.0 and use `data_disks` instead."
+  description = "(Deprecated) It has been deprecated from version 2.0.0 and use 'data_disks' instead."
   type        = number
   default     = 0
 }